Move expressions package (#26)
