+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/exactfraction/ExactFraction.kt
@@ -1,15 +1,15 @@
 package xyz.lbres.exactnumbers.exactfraction
 
-import xyz.lbres.common.CastingOverflowException
-import xyz.lbres.common.castToByte
-import xyz.lbres.common.castToChar
-import xyz.lbres.common.castToDouble
-import xyz.lbres.common.castToFloat
-import xyz.lbres.common.castToInt
-import xyz.lbres.common.castToLong
-import xyz.lbres.common.castToShort
-import xyz.lbres.common.createHashCode
-import xyz.lbres.common.divideByZero
+import xyz.lbres.exactnumbers.common.CastingOverflowException
+import xyz.lbres.exactnumbers.common.castToByte
+import xyz.lbres.exactnumbers.common.castToChar
+import xyz.lbres.exactnumbers.common.castToDouble
+import xyz.lbres.exactnumbers.common.castToFloat
+import xyz.lbres.exactnumbers.common.castToInt
+import xyz.lbres.exactnumbers.common.castToLong
+import xyz.lbres.exactnumbers.common.castToShort
+import xyz.lbres.exactnumbers.common.createHashCode
+import xyz.lbres.exactnumbers.common.divideByZero
 import xyz.lbres.exactnumbers.ext.eq
 import xyz.lbres.exactnumbers.ext.toExactFraction
 import xyz.lbres.kotlinutils.biginteger.ext.isNegative
