+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/irrationals/pi/Pi.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.exactnumbers.irrationals.pi
 
-import xyz.lbres.common.createHashCode
-import xyz.lbres.common.divideBigDecimals
+import xyz.lbres.exactnumbers.common.createHashCode
+import xyz.lbres.exactnumbers.common.divideBigDecimals
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumberCompanion
