+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/exactfraction/ExactFractionAlternateConstructorsTest.kt
@@ -1,7 +1,7 @@
 package xyz.lbres.exactnumbers.exactfraction
 
 import org.junit.Test
-import xyz.lbres.testutils.assertDivByZero
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import java.math.BigInteger
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
