+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/expressions/term/binaryOpsTests.kt
@@ -1,11 +1,11 @@
-package xyz.lbres.expressions.term
+package xyz.lbres.exactnumbers.expressions.term
 
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
-import xyz.lbres.testutils.TestNumber
-import xyz.lbres.testutils.assertDivByZero
+import xyz.lbres.exactnumbers.testutils.TestNumber
+import xyz.lbres.exactnumbers.testutils.assertDivByZero
 import kotlin.test.assertEquals
 
 private val logNum1 = Log(ExactFraction(15, 4))
